refactor: separate module is implemented to check that all requests to all routes, tested
