+++ score.md
@@ -24,7 +24,7 @@
 - [ ] **+10** `User` `password` saved into database as hash
 - [X] **+20** Access Token is implemented,`JWT` payload contains `userId` and `login`, secret key is saved in `.env`.
 - [X] **+40** Authentication is required for the access to all routes except `/auth/signup`, `/auth/login`, `/doc` and `/`.
-- [ ] **+10** Separate module is implemented **within application scope** to check that all requests to all routes except mentioned above contain required JWT token
+- [X] **+10** Separate module is implemented **within application scope** to check that all requests to all routes except mentioned above contain required JWT token
 
 ## Advanced Scope
